back_end/requirements.txt
beautifulsoup4==4.13.4
crawl4ai==0.6.1
duckduckgo_search==8.0.1
fastapi==0.115.12
ipython==8.12.3
ipython==8.30.0
langchain==0.3.24
langchain_community==0.3.22
langchain_core==0.3.56
langchain_openai==0.3.14
langgraph==0.3.34
nest_asyncio==1.6.0
pandas==2.2.3
Pillow==11.2.1
pydantic==2.11.3
python-dotenv==1.1.0
Requests==2.32.3
streamlit==1.42.1
uvicorn==0.34.2
